Update bitcoind.cpp
The added changes should fix the issue of the masternode.conf file not being able to be read / parsed when using masternode list-conf from the debug console. These changes were suggested by UjinM6 who is one of the main DASH devs which a lot of the GorillaBucks code is based on along with CRAVE.

https://dashtalk.org/threads/tao-suggested-i-post-here-0-05-btc-bounty-offered-for-resolution-to-my-masternode-issue.5704/

If changes are accepted, I would appreciate a new QT build,since it's no mahor changes it should I would think keep the same version number.